# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "A RESTful API for ratings, reviews, and comments system",
    "title": "Ratings and Reviews API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "API Support",
      "url": "http://www.example.com/support",
      "email": "support@example.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0"
  },
  "host": "localhost:8000",
  "basePath": "/api/v1",
  "paths": {
...
    "/ratings": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Create a new rating for a service",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "ratings"
        ],
        "summary": "Create a new rating",
        "parameters": [
          {
            "description": "Rating data",
            "name": "rating",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "service_id",
                "score"
              ],
              "properties": {
                "service_id": {
                  "type": "string",
                  "format": "uuid"
                },
                "score": {
                  "type": "integer",
                  "minimum": 1,
                  "maximum": 5
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Rating created successfully",
            "schema": {
              "type": "object"
            }
          },
          "400": {
            "description": "Invalid input",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/ratings/service/{serviceID}": {
      "get": {
        "description": "Retrieve all ratings for a specific service with pagination",
        "produces": [
          "application/json"
        ],
        "tags": [
          "ratings"
        ],
        "summary": "Get ratings for a service",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "description": "Service ID",
            "name": "serviceID",
            "in": "path",
            "required": true
          },
          {
            "type": "integer",
            "default": 10,
            "description": "Number of items per page",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "integer",
            "default": 0,
            "description": "Offset for pagination",
            "name": "offset",
            "in": "query"
          },
          {
            "type": "string",
            "default": "created_at",
            "description": "Field to sort by",
            "name": "sort_by",
            "in": "query"
          },
          {
            "enum": [
              "asc",
              "desc"
            ],
            "type": "string",
            "default": "desc",
            "description": "Sort direction",
            "name": "sort_direction",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "List of ratings with pagination metadata",
            "schema": {
              "type": "object",
              "properties": {
                "ratings": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                },
                "total": {
                  "type": "integer"
                },
                "limit": {
                  "type": "integer"
                },
                "offset": {
                  "type": "integer"
                }
              }
            }
          },
          "400": {
            "description": "Invalid service ID",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/ratings/service/{serviceID}/average": {
      "get": {
        "description": "Retrieve the average rating score for a specific service",
        "produces": [
          "application/json"
        ],
        "tags": [
          "ratings"
        ],
        "summary": "Get average rating for a service",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "description": "Service ID",
            "name": "serviceID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Average rating score",
            "schema": {
              "type": "number",
              "format": "float"
            }
          },
          "400": {
            "description": "Invalid service ID",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/ratings/service/{serviceID}/me": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Retrieve the authenticated user's rating for a specific service",
        "produces": [
          "application/json"
        ],
        "tags": [
          "ratings"
        ],
        "summary": "Get a user's rating for a service",
        "parameters": [
          {
            "type": "string",
            "format": "uuid",
            "description": "Service ID",
            "name": "serviceID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "User's rating for the service",
            "schema": {
              "type": "object"
            }
          },
          "400": {
            "description": "Invalid service ID",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Rating not found",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Type \"Bearer\" followed by a space and JWT token"
    }
  }
}